added actionsheet mock data to respective file
