--- src/app/core/test-data/my-expenses.page.spec.data.ts
@@ -1,26 +1,0 @@
-export const expectedActionSheetButtonRes = [
-  {
-    text: 'Capture Receipt',
-    icon: 'assets/svg/fy-camera.svg',
-    cssClass: 'capture-receipt',
-    handler: undefined,
-  },
-  {
-    text: 'Add Manually',
-    icon: 'assets/svg/fy-expense.svg',
-    cssClass: 'capture-receipt',
-    handler: undefined,
-  },
-  {
-    text: 'Add Mileage',
-    icon: 'assets/svg/fy-mileage.svg',
-    cssClass: 'capture-receipt',
-    handler: undefined,
-  },
-  {
-    text: 'Add Per Diem',
-    icon: 'assets/svg/fy-calendar.svg',
-    cssClass: 'capture-receipt',
-    handler: undefined,
-  },
-];
