+++ src/app/fyle/my-expenses/my-expenses.page.spec.ts
@@ -55,7 +55,7 @@ import { ExpenseFilters } from './expense-filters.model';
 import { txnData2 } from 'src/app/core/mock-data/transaction.data';
 import { unformattedTxnData } from 'src/app/core/mock-data/unformatted-transaction.data';
 import { expenseFiltersData1, expenseFiltersData2 } from 'src/app/core/mock-data/expense-filters.data';
-import { expectedActionSheetButtonRes } from 'src/app/core/test-data/my-expenses.page.spec.data';
+import { expectedActionSheetButtonRes } from 'src/app/core/mock-data/action-sheet-options.data';
 import { cloneDeep } from 'lodash';
 import { apiAuthRes } from 'src/app/core/mock-data/auth-reponse.data';
 import { cardDetailsData1 } from 'src/app/core/mock-data/card-details.data';
